% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_specificity.R
\name{calc_specificity}
\alias{calc_specificity}
\title{calc_specificity}
\usage{
calc_specificity(x, type = "tau")
}
\arguments{
\item{dataframe}{}
}
\value{
dataframe
}
\description{
calculates different specificity scores as defined by https://academic.oup.com/bib/article/18/2/205/2562739
}
\examples{
dataset <- data.frame(a = c(1,2,3), b = c(1,2,3), c = c(1,2,3))
normalized_data <- bioinfotool::calc_specificity(dataset, type='tau')


}
